PersonAuth dao fix

+++ lib-database/src/main/java/com/ustadmobile/core/db/dao/PersonAuthDao.java
@@ -52,6 +52,6 @@ public abstract class PersonAuthDao implements BaseDao<PersonAuth> {
 
 
     @UmUpdate
-    public abstract void updateAsync(PersonAuth entity, UmCallback<Long> resultObject);
+    public abstract void updateAsync(PersonAuth entity, UmCallback<Integer> resultObject);
 
 }
